sources/ni__measurement-plugin-service-extension-example/src/pyproject.toml
[tool.poetry]
name = "json_logger"
version = "1.0.0"
description = "A user-defined service to log the measurement data to a JSON file."
authors = ["NI <opensource@ni.com>"]

[tool.poetry.dependencies]
python = "^3.8"
ni-measurement-plugin-sdk-service = {version = "^2.0.0"}
click = ">=7.1.2, !=8.1.4" # mypy fails with click 8.1.4: https://github.com/pallets/click/issues/2558
grpcio = "*"
protobuf = "4.25.4"

[tool.poetry.group.dev.dependencies]
ni-python-styleguide = ">=0.4.1"
mypy = "^1.13.0"
grpc-stubs = "^1.53"
grpcio-tools = "1.59.0"
mypy-protobuf = "^3.6.0"
types-protobuf = "^5.28.3.20241030"
bandit = { version = "^1.7", extras = ["toml"] }

[tool.black]
extend_exclude = '\.tox/|_pb2(_grpc)?\.(py|pyi)$'
line-length = 100

[tool.ni-python-styleguide]
extend_exclude = '.tox/,*_pb2_grpc.py,*_pb2_grpc.pyi,*_pb2.py,*_pb2.pyi'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
ignore_missing_imports = true
exclude = ["^stubs/.*"]

[tool.bandit]
exclude_dirs = [".venv"]